# docs/swagger.yaml
basePath: /
definitions:
...
  service.AssignRoleRequest:
    properties:
      role_id:
        type: string
    type: object
  service.CreateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      role_id:
        type: string
      username:
        type: string
    type: object
...
  service.SetLecturerProfileRequest:
    properties:
      department:
        type: string
      lecturer_id:
        type: string
    type: object
  service.SetStudentProfileRequest:
    properties:
      academic_year:
        type: string
      advisor_id:
        type: string
      program_study:
        type: string
      student_id:
        type: string
    type: object
...
  service.UpdateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:4000
info:
  contact:
    email: support@example.com
    name: API Support
  description: API untuk sistem manajemen prestasi mahasiswa
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Achievement Management System API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get list of all users with pagination (Admin)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - User Management
    post:
      consumes:
      - application/json
      description: Create new user with role assignment (Admin)
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict - username/email exists
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create new user
      tags:
      - User Management
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user and associated profiles (Admin)
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - User Management
    get:
      consumes:
      - application/json
      description: Get detailed information of a user including profiles (Admin)
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user detail
      tags:
      - User Management
    put:
      consumes:
      - application/json
      description: Update user information (Admin)
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - User Management
  /api/v1/users/{id}/lecturer:
    post:
      consumes:
      - application/json
      description: Create or update lecturer profile (Admin)
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Lecturer profile data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/service.SetLecturerProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile saved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Set lecturer profile
      tags:
      - User Management
  /api/v1/users/{id}/role:
    put:
      consumes:
      - application/json
      description: Assign or change user role (Admin)
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Role data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/service.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role assigned
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Assign role to user
      tags:
      - User Management
  /api/v1/users/{id}/student:
    post:
      consumes:
      - application/json
      description: Create or update student profile with advisor (Admin)
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Student profile data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/service.SetStudentProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile saved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Set student profile
      tags:
      - User Management
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"